refactor(cli): extract init parser owner
